refactor: remove unused logging imports and clean up related code
