+++ packages/builder/lib/src/slide_processor.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 
-import 'package:logging/logging.dart';
 import 'package:superdeck_builder/src/parsers/raw_slide_schema.dart';
 import 'package:superdeck_core/superdeck_core.dart';
