--- packages/cli/example/superdeck_cli_example.dart
@@ -1,4 +1,0 @@
-void main() {
-  // Example usage of superdeck_cli
-  print('SuperDeck CLI Example');
-}
