+++ packages/cli/lib/src/commands/base_command.dart
@@ -2,14 +2,14 @@ import 'dart:convert';
 
 import 'package:args/command_runner.dart';
 import 'package:mason_logger/mason_logger.dart';
-import 'package:superdeck_core/superdeck_core.dart' hide logger;
+import 'package:superdeck_core/superdeck_core.dart' hide logger, Logger, Level;
 import 'package:yaml/yaml.dart';
 
-/// Base class for Superdeck commands with common functionality
-abstract class SuperdeckCommand extends Command<int> {
+/// Base class for SuperDeck commands with common functionality
+abstract class SuperDeckCommand extends Command<int> {
   final Logger logger;
 
-  SuperdeckCommand({Logger? loggerOverride})
+  SuperDeckCommand({Logger? loggerOverride})
     : logger = loggerOverride ?? Logger();
 
   /// Loads the Superdeck configuration from the default file
