+++ packages/cli/lib/src/utils/extensions.dart
@@ -1,6 +1,6 @@
 import 'package:args/command_runner.dart';
 
-extension CommandExtension on Command {
+extension CommandX on Command {
   /// Gets the parsed command-line option named [name] as `bool`.
   bool boolArg(String name) => argResults?[name] == true;
 }
