--- packages/core/example/superdeck_core_example.dart
@@ -1,3 +1,0 @@
-void main() {
-  // internal package for superdeck
-}
