+++ packages/core/lib/src/utils/extensions.dart
@@ -2,22 +2,8 @@ import 'dart:io';
 
 import 'package:ack/ack.dart';
 
-// List extensions
-extension ListX<T> on List<T> {
-  /// Returns a list of items that have been added and removed between this list and another
-  ({List<T> added, List<T> removed}) compareWith(List<T> other) {
-    final selfSet = toSet();
-    final otherSet = other.toSet();
-
-    return (
-      added: other.where((item) => !selfSet.contains(item)).toList(),
-      removed: where((item) => !otherSet.contains(item)).toList(),
-    );
-  }
-}
-
 // File extensions
-extension FileExt on File {
+extension FileX on File {
   /// Ensures a file exists and optionally writes content to it
   Future<File> ensureExists({String? content}) async {
     if (!await exists()) {
@@ -39,7 +25,7 @@ extension FileExt on File {
 }
 
 // Directory extensions
-extension DirectoryExt on Directory {
+extension DirectoryX on Directory {
   /// Ensures a directory exists, creating it if necessary
   Future<Directory> ensureExists() async {
     if (!await exists()) {
